retire deprecated code
