+++ pyod/utils/__init__.py
@@ -8,7 +8,6 @@ from .utility import argmaxn
 from .utility import invert_order
 from .data import generate_data
 from .data import evaluate_print
-from .data import get_color_codes
 from .stat_models import pairwise_distances_no_broadcast
 from .stat_models import wpearsonr
 from .stat_models import pearsonr_mat
@@ -22,7 +21,6 @@ __all__ = ['check_parameter',
            'invert_order',
            'generate_data',
            'evaluate_print',
-           'get_color_codes',
            'pairwise_distances_no_broadcast',
            'wpearsonr',
            'pearsonr_mat']
